requirements.txt
fastapi
uvicorn
feedparser
sqlalchemy
aiohttp
google-genai
python-dotenv
beautifulsoup4
python-multipart
